Organize core utilities
