+++ core/structured_logger.py
